test: improve logs, move fns into utils
