+++ tests/swap.rs
@@ -18,7 +18,6 @@ use std::str::FromStr;
 
 use ntest::timeout;
 
-use utils::config;
 use utils::fc::*;
 use utils::setup_logging;
 
@@ -2581,14 +2580,6 @@ fn kill_connected_peerd() {
     .expect("Sending CTR-C to peerd failed");
 }
 
-fn reusable_btc_address() -> bitcoin::Address {
-    bitcoin::Address::from_str("bcrt1q3rc4sm3w9fr6a46n08znfjt7eu2yhhel6j8rsa").unwrap()
-}
-
-fn reusable_xmr_address() -> monero::Address {
-    monero::Address::from_str("44CpGC77Kn6exUWYCUwfaUYmDeKn7MyRcNPikgeHBCz8M6LXUC3fGCWNMW7UACHyTL6QxzqKxvJbu5o2VESLzCaeNHNUkwv").unwrap()
-}
-
 fn info_args(data_dir: Vec<String>) -> Vec<String> {
     data_dir
         .into_iter()
@@ -2887,7 +2878,7 @@ async fn retry_until_bitcoin_funding_address(
         let funding_infos: Vec<BitcoinFundingInfo> = stdout
             .iter()
             .filter_map(|element| {
-                info!("element: {:?}", element);
+                info!("cli: {}", element);
                 if let Ok(funding_info) = BitcoinFundingInfo::from_str(element) {
                     if funding_info.swap_id == swap_id {
                         Some(funding_info)
@@ -2916,7 +2907,7 @@ async fn retry_until_funding_info_cleared(swap_id: SwapId, args: Vec<String>) {
         let funding_infos: Vec<BitcoinFundingInfo> = stdout
             .iter()
             .filter_map(|element| {
-                info!("element: {:?}", element);
+                info!("cli: {}", element);
                 if let Ok(funding_info) = BitcoinFundingInfo::from_str(element) {
                     if funding_info.swap_id == swap_id {
                         Some(funding_info)
@@ -2948,6 +2939,7 @@ async fn retry_until_monero_funding_address(
         let funding_infos: Vec<MoneroFundingInfo> = stdout
             .iter()
             .filter_map(|element| {
+                info!("cli: {}", element);
                 if let Ok(funding_info) = MoneroFundingInfo::from_str(element) {
                     if funding_info.swap_id == swap_id {
                         Some(funding_info)
@@ -2979,6 +2971,7 @@ async fn retry_until_bob_finish_state_transition(
         let bob_finish: Vec<String> = stdout
             .iter()
             .filter_map(|element| {
+                info!("cli: {}", element);
                 if element.contains(&finish_state) {
                     Some(element.to_string())
                 } else {
@@ -3014,7 +3007,7 @@ async fn retry_until_finish_state_transition(
         let alice_finish: Vec<String> = stdout
             .iter()
             .filter_map(|element| {
-                info!("element: {:?}", element);
+                info!("cli: {}", element);
                 if element.contains(&finish_state) {
                     Some(element.to_string())
                 } else {
@@ -3033,95 +3026,4 @@ async fn retry_until_finish_state_transition(
         "timeout before finish state {:?} could be retrieved",
         finish_state
     );
-}
-
-async fn monero_setup() -> (
-    monero_rpc::RegtestDaemonJsonRpcClient,
-    Arc<Mutex<monero_rpc::WalletClient>>,
-) {
-    let conf = config::TestConfig::parse();
-    let client = monero_rpc::RpcClient::new(format!("{}", conf.monero.daemon));
-    let regtest = client.daemon().regtest();
-    let client = monero_rpc::RpcClient::new(format!(
-        "{}",
-        conf.monero.get_wallet(config::WalletIndex::Primary)
-    ));
-    let wallet = client.wallet();
-
-    // error happens when wallet does not exist
-    if wallet.open_wallet("test".to_string(), None).await.is_err() {
-        wallet
-            .create_wallet("test".to_string(), None, "English".to_string())
-            .await
-            .unwrap();
-        wallet.open_wallet("test".to_string(), None).await.unwrap();
-    }
-
-    let address = wallet.get_address(0, None).await.unwrap();
-    regtest.generate_blocks(200, address.address).await.unwrap();
-    regtest
-        .generate_blocks(200, reusable_xmr_address())
-        .await
-        .unwrap();
-
-    (regtest, Arc::new(Mutex::new(wallet)))
-}
-
-async fn monero_new_dest_address(
-    wallet: Arc<Mutex<monero_rpc::WalletClient>>,
-) -> (monero::Address, String) {
-    use rand::distributions::Alphanumeric;
-    use rand::{thread_rng, Rng};
-
-    let wallet_name: String = thread_rng()
-        .sample_iter(&Alphanumeric)
-        .take(15)
-        .map(char::from)
-        .collect();
-
-    let wallet_lock = wallet.lock().await;
-
-    wallet_lock
-        .create_wallet(wallet_name.clone(), None, "English".to_string())
-        .await
-        .unwrap();
-    wallet_lock
-        .open_wallet(wallet_name.clone(), None)
-        .await
-        .unwrap();
-    let address = wallet_lock.get_address(0, None).await.unwrap();
-
-    (address.address, wallet_name)
-}
-
-async fn send_monero(
-    wallet: Arc<Mutex<monero_rpc::WalletClient>>,
-    address: monero::Address,
-    amount: monero::Amount,
-) -> Vec<u8> {
-    let options = monero_rpc::TransferOptions {
-        account_index: None,
-        subaddr_indices: None,
-        mixin: None,
-        ring_size: None,
-        unlock_time: None,
-        payment_id: None,
-        do_not_relay: None,
-    };
-    let wallet_lock = wallet.lock().await;
-    wallet_lock
-        .open_wallet("test".to_string(), None)
-        .await
-        .unwrap();
-    wallet_lock.refresh(Some(1)).await.unwrap();
-    let transaction = wallet_lock
-        .transfer(
-            [(address, amount)].iter().cloned().collect(),
-            monero_rpc::TransferPriority::Default,
-            options,
-        )
-        .await
-        .unwrap();
-    drop(wallet_lock);
-    hex::decode(transaction.tx_hash.to_string()).unwrap()
 }
